Moved board to its own file
